Split logger plugins into individual files; renamed multiplicative evaluator file
